Check that Fortran module is created  - Looks like Absoft has problems
Hg-commit: ed6b5a321ffe974752c1ce150461b65ccb9fd9e8

+++ config/compilers.py
@@ -1103,6 +1103,8 @@ class Configure(config.base.Configure):
     if not os.path.isdir(testdir):
       os.mkdir(testdir)
     os.rename(self.compilerObj, modobj)
+    if not os.path.isfile(modname):
+      raise RuntimeError('Fortran module '+modname+' was not created during the compile')
     os.rename(modname, os.path.join(testdir, modname))
     fcode = '''\
       use configtest
